moved Environment and World to subfolder of server side
