+++ src/main/java/com/github/migi_1/Context/App.java
@@ -1,5 +1,6 @@
 package com.github.migi_1.Context;
 
+import com.github.migi_1.Context.serverside.model.World;
 import com.jme3.app.SimpleApplication;
 import com.jme3.renderer.RenderManager;
